+++ apps/docs/src/components/document/resources/download-icon-button.tsx
@@ -1,22 +1,41 @@
 'use client';
 import { IconButton } from '@ogcio/design-system-react';
+import type { IconProps } from '@ogcio/design-system-react/icons';
 import { useEffect, useState } from 'react';
+import type { ComponentType } from 'react';
 import { cn } from '@/lib/cn';
 import analytics from '@/lib/analytics';
+import { CheckIcon, DownloadIcon } from '@ogcio/design-system-react/icons';
+import { createRoot } from 'react-dom/client';
+import { flushSync } from 'react-dom';
 
 type SvgIcon = {
   name: string;
   href: string;
   text?: never;
+  IconComponent?: never;
 };
 
 type TextIcon = {
   name: string;
   text: string;
   href?: never;
+  IconComponent?: never;
 };
 
-export function DownloadIconButton({ name, href, text }: SvgIcon | TextIcon) {
+type SourceIcon = {
+  IconComponent: ComponentType<IconProps>;
+  name: string;
+  href?: never;
+  text?: never;
+};
+
+export function DownloadIconButton({
+  name,
+  href,
+  text,
+  IconComponent,
+}: SvgIcon | TextIcon | SourceIcon) {
   const [downloaded, setDownloaded] = useState(false);
 
   useEffect(() => {
@@ -28,9 +47,18 @@ export function DownloadIconButton({ name, href, text }: SvgIcon | TextIcon) {
     return () => clearTimeout(timeout);
   }, [downloaded]);
 
-  const safeName = name.toLowerCase().trim().replace(/\s+/g, '_');
-
   const handleClick = () => {
+    const link = document.createElement('a');
+    link.download = `${name}.svg`;
+    if (text) {
+      link.href = encodedSvgURL(text);
+    } else if (href) {
+      link.href = href;
+    } else if (IconComponent !== undefined) {
+      link.href = getSvgUri(IconComponent);
+    }
+    link.click();
+
     analytics.trackEvent({
       category: 'download content',
       action: 'click',
@@ -38,22 +66,35 @@ export function DownloadIconButton({ name, href, text }: SvgIcon | TextIcon) {
     });
     setDownloaded(true);
   };
-
-  const downloadHref = text
-    ? `data:image/svg+xml;charset=utf-8,${encodeURIComponent(text)}`
-    : href || '';
-
   return (
-    <a href={downloadHref} download={`${safeName}.svg`} onClick={handleClick}>
-      <IconButton
-        icon={{ icon: downloaded ? 'check' : 'download' }}
-        size="small"
-        appearance="light"
-        className={cn({
-          'text-green-600': downloaded,
-          'text-gray-600': !downloaded,
-        })}
-      />
-    </a>
+    <IconButton
+      onClick={handleClick}
+      size="sm"
+      appearance="light"
+      className={cn({
+        'text-green-600': downloaded,
+        'text-gray-600': !downloaded,
+      })}
+    >
+      {downloaded ? (
+        <CheckIcon label="downloaded" />
+      ) : (
+        <DownloadIcon label="download" />
+      )}
+    </IconButton>
   );
 }
+
+const encodedSvgURL = (text: string) =>
+  `data:image/svg+xml;charset=utf-8,${encodeURIComponent(text)}`;
+
+const getSvgUri = (Icon: ComponentType<IconProps>) => {
+  const iconContainer = document.createElement('div');
+  const root = createRoot(iconContainer);
+  flushSync(() => {
+    root.render(<Icon size={48} />);
+  });
+  const svgUrl = encodedSvgURL(iconContainer.innerHTML);
+  root.unmount();
+  return svgUrl;
+};